Initial Commit
The ML code has been created with a scalable format for a CNN. It uses the Adam optimizer and a LR scheduler as well.
It also has support for early stopping and has a helper function to display a batch.

The blender python file has an input parameter dictionary that can be set to generate images of varying requirements.